Major Updates
GM - Updated Yusra's python code to change color of images. Images are now being deployed on the respective html files. These pages should now run locally. I also added content and updated the names of the pages and the links.
I did not make any changes to the app.py or to the calls.